Separate wallet probe backend service
